+++ CHANGES.md
@@ -5,6 +5,7 @@
 -   Release all artifacts to GitHub Container Registry (instead of docker.io & https://charts.magda.io)
 -   Upgrade magda-common chart version
 -   Upgrade api to batch/v1 to be compatible with k8s 1.25 (now requires >=1.21)
+-   Use node-fetch for http request instead
 
 # 1.0.0